add launcher switch and attempt to add launcher to docker-images

+++ packages/cli/src/runners/task-runner-process.ts
@@ -38,28 +38,44 @@ export class TaskRunnerProcess {
 		a.ok(!this.process, 'Task Runner Process already running');
 
 		const grantToken = await this.authService.createGrantToken();
-		// const startScript = require.resolve('@n8n/task-runner');
 
-		// this.process = spawn('node', [startScript], {
-		// 	env: {
-		// 		PATH: process.env.PATH,
-		// 		N8N_RUNNERS_GRANT_TOKEN: grantToken,
-		// 		N8N_RUNNERS_N8N_URI: `localhost:${this.globalConfig.port}`,
-		// 	},
-		// });
-
-		this.process = spawn(this.globalConfig.taskRunners.launcherPath, ['launch', 'javascript'], {
-			env: {
-				PATH: process.env.PATH,
-				N8N_RUNNERS_GRANT_TOKEN: grantToken,
-				N8N_RUNNERS_N8N_URI: `127.0.0.1:${this.globalConfig.taskRunners.port}`,
-			},
-		});
+		const n8nUri = `127.0.0.1:${this.globalConfig.taskRunners.port}`;
+		this.process = this.globalConfig.taskRunners.useLauncher
+			? this.startLauncher(grantToken, n8nUri)
+			: this.startNode(grantToken, n8nUri);
 
 		this.process.stdout?.pipe(process.stdout);
 		this.process.stderr?.pipe(process.stderr);
 
 		this.monitorProcess(this.process);
+	}
+
+	startNode(grantToken: string, n8nUri: string) {
+		const startScript = require.resolve('@n8n/task-runner');
+
+		return spawn('node', [startScript], {
+			env: {
+				PATH: process.env.PATH,
+				N8N_RUNNERS_GRANT_TOKEN: grantToken,
+				N8N_RUNNERS_N8N_URI: n8nUri,
+			},
+		});
+	}
+
+	startLauncher(grantToken: string, n8nUri: string) {
+		return spawn(
+			this.globalConfig.taskRunners.launcherPath,
+			['launch', this.globalConfig.taskRunners.launcherRunner],
+			{
+				env: {
+					PATH: process.env.PATH,
+					N8N_RUNNERS_GRANT_TOKEN: grantToken,
+					N8N_RUNNERS_N8N_URI: n8nUri,
+					// For debug logging if enabled
+					RUST_LOG: process.env.RUST_LOG,
+				},
+			},
+		);
 	}
 
 	@OnShutdown()
@@ -71,10 +87,39 @@ export class TaskRunnerProcess {
 		this.isShuttingDown = true;
 
 		// TODO: Timeout & force kill
-		this.process.kill();
+		if (this.globalConfig.taskRunners.useLauncher) {
+			await this.killLauncher();
+		} else {
+			this.killNode();
+		}
 		await this.runPromise;
 
 		this.isShuttingDown = false;
+	}
+
+	killNode() {
+		if (!this.process) {
+			return;
+		}
+		this.process.kill();
+	}
+
+	async killLauncher() {
+		if (!this.process?.pid) {
+			return;
+		}
+
+		const process = spawn(this.globalConfig.taskRunners.launcherPath, [
+			'kill',
+			'javascript',
+			this.process.pid.toString(),
+		]);
+
+		await new Promise<void>((resolve) => {
+			process.on('exit', () => {
+				resolve();
+			});
+		});
 	}
 
 	private monitorProcess(process: ChildProcess) {
